+++ openelis/src/us/mn/state/health/lims/systemusermodule/daoimpl/RoleModuleDAOImpl.java
@@ -17,15 +17,9 @@
 */
 package us.mn.state.health.lims.systemusermodule.daoimpl;
 
-import java.math.BigInteger;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Vector;
-
 import org.apache.commons.beanutils.PropertyUtils;
 import org.hibernate.HibernateException;
 import org.hibernate.Query;
-
 import us.mn.state.health.lims.audittrail.dao.AuditTrailDAO;
 import us.mn.state.health.lims.audittrail.daoimpl.AuditTrailDAOImpl;
 import us.mn.state.health.lims.common.action.IActionConstants;
@@ -41,6 +35,10 @@ import us.mn.state.health.lims.systemusermodule.valueholder.PermissionModule;
 import us.mn.state.health.lims.systemusermodule.valueholder.RoleModule;
 import us.mn.state.health.lims.userrole.dao.UserRoleDAO;
 import us.mn.state.health.lims.userrole.daoimpl.UserRoleDAOImpl;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Vector;
 
 /**
  *  
@@ -336,7 +334,7 @@ public class RoleModuleDAOImpl extends BaseDAOImpl implements PermissionAgentMod
 			String sql = "select count(*) from system_user_role where system_user_id = :userId";
 			Query query = HibernateUtil.getSession().createSQLQuery(sql);
 			query.setInteger("userId", userId);
-			int roleCount = ((BigInteger)query.uniqueResult()).intValue();
+			int roleCount = ((Long)query.uniqueResult()).intValue();
 			return roleCount > 0;
 		}catch( HibernateException he){
 			LogEvent.logError("RoleModuleDAOImpl","doesUserHaveAnyModules(int)",he.toString());
